Ported GRU and GRU dropout test from CTest to GTest
